Implement vehicle locations on the front end (#54)
* Added vehicle locations on the map view

* Reset route selection in StopsDetails component on stopDetails change

In the StopsDetails component, added a useEffect hook to reset selectedRoute, selectedDirection, and selectedRouteTag to null when the stopDetails prop changes. This ensures a clean state when a new stop is selected, allowing users to choose a route and direction as if it's the first selection.

* Set the icon of the stops the default google maps marker icon

* Created the controller, view and data files for the vehicle_locations use-case

* Added vehicle locations on the map view

* Reset route selection in StopsDetails component on stopDetails change

In the StopsDetails component, added a useEffect hook to reset selectedRoute, selectedDirection, and selectedRouteTag to null when the stopDetails prop changes. This ensures a clean state when a new stop is selected, allowing users to choose a route and direction as if it's the first selection.

* Set the icon of the stops the default google maps marker icon

* Created the controller, view and data files for the vehicle_locations use-case