Updated VehicleLocationsController to take in additional dirTag parameter. (#61)
Updated VehicleLocationsService to take in additional dirTag parameter.

Updated VehicleLocationsImplementation to filter vehicles on a route by their direction tag.

Added Tests for Vehicle Locations use case. Tests check for cases where the implementation returns a nonempty list of Vehicles, an empty list of Vehicles due to no vehicles existing with the given dirTag, an empty list of Vehicles due to no vehicles existing for a given routeTag, and an empty list of Vehicles due to when the routeTag does not exist.